requirements/development.txt
-r base.txt

django-debug-toolbar
psycopg[binary]
